fix: resolve compilation errors in build
- Fix return type mismatch in PerformanceStatistics.kt
- Add coroutineContext import in StreamingOptimizationManager.kt
- Add fillMaxWidth import in TopologyScreen.kt
- Fix strokeWidth shadowing in NetworkGraphCanvas.kt
- Use getApplication() instead of private application property in GamingViewModel.kt
- Fix coroutine scope issues in StreamingViewModel.kt

+++ app/src/main/kotlin/com/simplexray/an/performance/statistics/PerformanceStatistics.kt
@@ -46,7 +46,7 @@ class PerformanceStatistics {
         val rank = (percentile / 100.0) * (sortedValues.size - 1)
         val lower = rank.toInt()
         val upper = lower + 1
-        val fraction = rank - lower
+        val fraction = (rank - lower).toFloat()
 
         return if (upper >= sortedValues.size) {
             sortedValues.last()
